requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
openpyxl
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
